// frontend/src/api/model/permission.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Hospital Internment API
 * OpenAPI spec version: 0.1.0
 */

export type Permission = typeof Permission[keyof typeof Permission];


export const Permission = {
  ADMISSION: 'ADMISSION',
  HOSPITALIZATION: 'HOSPITALIZATION',
  BED_CLEANING: 'BED_CLEANING',
  NURSING_TASKS: 'NURSING_TASKS',
  BED_MANAGEMENT: 'BED_MANAGEMENT',
  BILLING: 'BILLING',
  CATALOG: 'CATALOG',
  USER_ADMIN: 'USER_ADMIN',
} as const;